pyproject.toml
[tool.poetry]
name = "weekend_fun"
version = "0.1.0"
description = "AI-powered event recommendation system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.12.0"
jina = "^3.15.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
sendgrid = "^6.10.0"
setuptools = "^69.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.10.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 